feat: #224 - EditRoutineVC에서 EditExerciseVC로 화면전환 역할 Coordinator로 분류
